Problem: `jetstream.allocation.TASAPIDriver#get_tacc_username` returns the Atmosphere username when it fails to get a valid TACC username.
This has caused situations where we reported allocation usage back to TAS with an invalid TACC username.

Solution: Re-raise the exception instead of silently swallowing it, and don't save the invalid TACC username into the `username_map`

Note: This is going to have downstream effects, particularly when creating reports. We should handle this problem there.